requirements.txt
pyzabbix
pandas
matplotlib
fpdf2
python-dotenv
requests
streamlit